+++ CHANGES.md
@@ -2,4 +2,4 @@
 =======
 
 * automatically load Customer Token from the URL if not available in a cookie
-
+* coupon campaign methods added.